refactor src.vex._abstract_device
