+++ src/vex/brain/screen.py
@@ -5,7 +5,7 @@ from collections.abc import Sequence
 
 from abm.decor import act
 
-from .._abstract import SingletonDevice
+from .._abstract_device import SingletonDevice
 from ..util.doc import robotmesh_doc, vexcode_doc
