+++ src/vex/bumper_switch_sensor.py
@@ -5,8 +5,8 @@ from collections.abc import Sequence
 
 from abm.decor import sense
 
-from ._abstract import Device
-from .port import Ports
+from ._abstract_device import Device
+from .brain.port import Ports
 
 # pylint: disable=unused-import
 from .util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
@@ -15,6 +15,9 @@ from .util.doc import robotmesh_doc, vexcode_doc   # noqa: F401
 __all__: Sequence[str] = ('Bumper',)
 
 
+@robotmesh_doc("""
+    robotmesh.com/studio/content/docs/vexiq-python_b/html/classvex_1_1_bumper.html
+""")
 class Bumper(Device):
     """Bumper Switch Sensor."""
