+++ src/vex/motor/__init__.py
@@ -7,8 +7,8 @@ from typing_extensions import Self
 
 from abm.decor import act, sense
 
-from .._abstract import Device
-from ..port import Ports
+from .._abstract_device import Device
+from ..brain.port import Ports
 from ..time import TimeUnits
 from ..units_common import RotationUnits
