+++ src/vex/units_common/distance_units.py
@@ -13,6 +13,8 @@ __all__: Sequence[str] = ('DistanceUnits',)
 
 @robotmesh_doc("""
     The measurement units for distance values.
+
+    robotmesh.com/studio/content/docs/vexiq-python_b/html/classvex_1_1_distance_units.html
 """)
 class DistanceUnits(IntEnum):
     """Distance Units."""
